+++ Wrappers/Python/ccpi/viewer/viewerLinker.py
@@ -3,6 +3,8 @@
 from ccpi.viewer.CILViewer import CILInteractorStyle as CIL3DInteractorStyle
 from ccpi.viewer.CILViewer2D import CILInteractorStyle as CIL2DInteractorStyle
 
+from ccpi.viewer.widgets.slider import SLIDER_EVENT
+
 
 class Linked3DInteractorStyle(CIL3DInteractorStyle):
     """
@@ -310,6 +312,19 @@ class ViewerLinkObserver():
                 # Linked, check if orientation is the same
                 if (self.sourceVtkViewer.getSliceOrientation() != self.targetVtkViewer.getSliceOrientation()):
                     shouldPassEvent = False
+
+        # Slice from the slider
+        if event == "NoEvent" and self.linkSlice:
+            # Although I invoked a SLIDER_EVENT in widgets/slider.py, I noticed that a
+            # NoEvent is actually emitted/received. Therefore, we will link the slice in case of a
+            # NoEvent and self.linkSlice=True
+            if (not self.linkSlice):
+                shouldPassEvent = False
+            else:
+                # Set current slice
+                sliceno = self.sourceViewer.getActiveSlice()
+                self.targetInteractor.GetInteractorStyle().SetActiveSlice(sliceno)
+                self.targetInteractor.GetInteractorStyle().UpdatePipeline(True)
 
         # KeyPress and orientation
         if (event == "KeyPressEvent" or state == 1026):